% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCellCyclePhase.R
\docType{methods}
\name{addCellCyclePhase}
\alias{addCellCyclePhase}
\alias{addCellCyclePhase,BioData-method}
\title{description of function addCellCyclePhase}
\usage{
addCellCyclePhase(x, s.genes = NULL, g2m.genes = NULL, gnameCol = NULL)
}
\arguments{
\item{x}{the BioData object}

\item{s.genes}{the s genes (defualt NULL means Seurat::cc.genes$s.genes )}

\item{g2m.genes}{the s genes (defualt NULL means Seurat::cc.genes$g2m.genes )}

\item{gnameCol}{the name of the Gene Symbol column in the annotation table (default NULL)}
}
\description{
Uses Seurat to address the cell cycle phase
}
