% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\docType{methods}
\name{as.BioData}
\alias{as.BioData}
\title{description of function as.BioData}
\usage{
as.BioData(dat)
}
\arguments{
\item{dat}{the counts list you get from Rsubread::featureCounts()}
}
\description{
create a NGSexpressionSet from a counts list object
}
