% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BioData.R
\docType{methods}
\name{as_BioData}
\alias{as_BioData}
\alias{as_BioData,list-method}
\alias{as_BioData,matrix-method}
\alias{as_BioData,data.frame-method}
\alias{as_BioData,cellexalvrR-method}
\alias{as_BioData,Seurat-method}
\alias{as_BioData,character-method}
\title{create a BioData from an other object}
\usage{
as_BioData(dat, ...)

\S4method{as_BioData}{list}(dat)

\S4method{as_BioData}{matrix}(dat)

\S4method{as_BioData}{data.frame}(dat)

\S4method{as_BioData}{cellexalvrR}(dat)

\S4method{as_BioData}{Seurat}(dat)

\S4method{as_BioData}{character}(dat, minUMI = 100, minGexpr = NULL)
}
\arguments{
\item{dat}{a sqlite database as created by the cellexalvrR::export2cellexalvr() function}

\item{...}{source object specififc options}

\item{minUMI}{the minimum UMI count in order to read a cell (default=100)}

\item{minGexpr}{the minimum UMI count for a gene to be included (default NULL - translates into 1)}
}
\description{
create a BioData from an other object

Supports sqlite3 databases as used in the cellexalVR application, 
10X output files and kallisto out files.
}
\section{Functions}{
\itemize{
\item \code{as_BioData,list-method}: create a BioData::R6 object from a Rsubread result list

\item \code{as_BioData,matrix-method}: convert a matrix to a BioData::R6 object

\item \code{as_BioData,data.frame-method}: convert a data.frame to a BioData::R6 object

\item \code{as_BioData,cellexalvrR-method}: convert a cellexalvrR object to a SingleCells::BioData::R6 object

\item \code{as_BioData,Seurat-method}: convert a seurat object into a SingleCells::BioData::R6 object

\item \code{as_BioData,character-method}: character path or sqlite3 database name
}}

