% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_order_grouping.R
\docType{methods}
\name{auto_order_grouping}
\alias{auto_order_grouping}
\alias{auto_order_grouping,BioData-method}
\title{Auto re-order a (RF) grouping}
\usage{
auto_order_grouping(x, group, settings = list(A = "lsens2017-3-2", t =
  "01:00:00", p = "dell", n = 1, N = 1), k = 5, colapseBy = "median")
}
\arguments{
\item{x}{The BioData object}

\item{group}{the sample grouping that should become reordered (min 15 groups)}

\item{settings}{a list of slurm settings default}

\item{k}{how many groups to put into one super group (default 5)}

\item{colapseBy}{which method to collapse the data by (default 'median') see function collaps
list( 'A' = 'lsens2017-3-2' , 't'= '01:00:00', 'p' = 'dell', 'n' = 1, 'N' = 1)}
}
\description{
This function uses the rfCluster_col function to order median collapsed data values.
This RF order is then re-ordered twice using hclust 'ward.D2' and the final ordering is forced onto the original grouping.
}
