% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_reorder_grouping.R
\docType{methods}
\name{auto_reorder_grouping}
\alias{auto_reorder_grouping}
\alias{auto_reorder_grouping,BioData-method}
\title{automaticly reorder a grouping based on summmary expression hclust}
\arguments{
\item{x}{the SingleCells object}

\item{group}{The grouping you wat to reorder}
}
\description{
reorder the grouping of single cell data creating assumed summary data
}
