% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestGrouping.R
\docType{methods}
\name{bestGrouping}
\alias{bestGrouping}
\alias{'bestGrouping,BioData-method}
\title{description of function randomForest}
\arguments{
\item{x}{the single cells ngs object}

\item{group}{a vector of sample columns that should be checked (the most complex is used only)}

\item{bestColname}{the column name to store the best grouping in}

\item{cutoff}{the cutoff percentage where all groups showing less than this percentacge of remapped samples are dropped}

\item{ntree}{for the rf process (default 2000)}

\item{...}{additional variables for the randomForest call}
}
\value{
a distRF object to be analyzed by pamNew
}
\description{
The function is using a randomForest classifier with 2000 trees to classify the given data using the given grooping

All groups that fail to be prediceted using the random forest are deemed ungrouped.

All groups where less than 50 percent of the total samples geting classified as being from that group fail.
}
\examples{
## lets data be a source BioData object and selection be a smaller one with only selected samples 
## for which a grouping has been produced named 'new_perfect_grouping'
## Not run:
predictor <- bestGrouping( selection, 'new_perfect_grouping' )
data$samples$new_perfect_grouping <- predict(predictor, as.matrix(t(data$dat)))
## End(**Not run**)
}
