% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata-package.R
\docType{package}
\name{biodata-package}
\alias{biodata}
\alias{biodata-package}
\title{biodata: Example datasets and analyses from the life sciences.}
\description{
biodata is a repository of datasets from throughout the life sciences which can be used for demonstrating, exploring, and learning data science, statistic, and machine learning.
}
\author{
\strong{Maintainer}: Nathan Brouwer \email{brouwernt@gmail.com}

}
\keyword{internal}
