% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class.BioData.R
\docType{class}
\name{BioData}
\alias{BioData}
\title{Class a simple interface to biological data (numeric) and rich annotation for both columns (samples) and rows (values)}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}} to store BioData.
}
\description{
Class a simple interface to biological data (numeric) and rich annotation for both columns (samples) and rows (values)

Class a simple interface to biological data (numeric) and rich annotation for both columns (samples) and rows (values)
}
\examples{
set.seed(1)
dat = data.frame( matrix(rnorm(1000),ncol=10) ) 
colnames(dat) <- paste('Sample', 1:10)
rownames(dat) <- paste( 'gene', 1:100)
samples <- data.frame(SampleID = 1:10, sname = colnames(dat) )
annotation <- data.frame( GeneID = paste( 'gene', 1:100), Start= 101:200 )
x <- BioData$new( cbind(annotation,dat), 
	Samples=samples, name="testObject",namecol='sname', outpath = "" )
}
\keyword{BioData}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{the numerical data as sparse Matrix}

\item{\code{raw}}{the raw numerical data as sparse Matrix}

\item{\code{raw_intronic}}{the raw numerical intronic data as sparse Matrix}

\item{\code{samples}}{the sample annotation as data.frame}

\item{\code{annotation}}{the row annotation as data.frame}

\item{\code{usedObj}}{a multi purpose list to store whichever ananlyis results do not fit in the stats list}

\item{\code{stats}}{all stats with one result for each data row}

\item{\code{ranks}}{currently unused}

\item{\code{logged}}{a logical value stating if the data has been logged}

\item{\code{snorm}}{a logical value stating if the object has been sample or cell normalized}

\item{\code{rownamescol}}{which column in the annotation table contains the data rownames}

\item{\code{sampleNamesCol}}{which column in the samples table contains the data rownames}

\item{\code{outpath}}{where to store the object}

\item{\code{name}}{the name of the object (e.g. project name)}

\item{\code{drop=c('MDS')}}{currently unused}

\item{\code{version}}{the version string of the BioData library that created this object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{the numerical data as sparse Matrix}

\item{\code{drop=c('MDS')}}{currently unused}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{BioData$print()}}
\item \href{#method-new}{\code{BioData$new()}}
\item \href{#method-data}{\code{BioData$data()}}
\item \href{#method-rawData}{\code{BioData$rawData()}}
\item \href{#method-force.numeric}{\code{BioData$force.numeric()}}
\item \href{#method-pwd}{\code{BioData$pwd()}}
\item \href{#method-forceAbsoluteUniqueSample}{\code{BioData$forceAbsoluteUniqueSample()}}
\item \href{#method-clone}{\code{BioData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print a summary of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
initialize the object - depricated - use the as_BioData functions instead
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$new(
  dat,
  Samples,
  annotation = NULL,
  name = "BioData",
  namecol = NULL,
  namerow = "GeneID",
  outpath = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{the expression sparse Matrix}

\item{\code{Samples}}{a data.frame describing the column data}

\item{\code{annotation}}{a data.frame describing the row data}

\item{\code{name}}{the name of the project - make it phony - will be used as filename}

\item{\code{namecol}}{the column in the Samples data that is the rownames of the data Matrix}

\item{\code{namerow}}{the column in the annotation data that is the rownames of the data Matrix}

\item{\code{outpath}}{the path the file is stored in (defaults to getwd())}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\subsection{Method \code{data()}}{
data accessor function either reports the dat spot or the zscored if existing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rawData"></a>}}
\subsection{Method \code{rawData()}}{
data accessor function either reports the dat spot or the raw if existing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$rawData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-force.numeric"></a>}}
\subsection{Method \code{force.numeric()}}{
useless and depricated
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$force.numeric()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pwd"></a>}}
\subsection{Method \code{pwd()}}{
useless
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$pwd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forceAbsoluteUniqueSample"></a>}}
\subsection{Method \code{forceAbsoluteUniqueSample()}}{
likely useless and sotherwise implemented, but makes sure all entries in the vector are unique.
by default adds _1, _2, ... , _n to the not unique names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$forceAbsoluteUniqueSample(x, separator = "_")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the vector}

\item{\code{separator}}{by default '_'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
