% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_files.R
\docType{data}
\name{CellCycleGenes}
\alias{CellCycleGenes}
\title{A simple table containing the Human and mouse orthologe ensembl ids for the cell cycle genes used in PMID28263960 using BioMart and GRCh38.p10}
\format{data.frame}
\usage{
CellCycleGenes
}
\description{
The BioMart web interface was used to create the input data, which in turn was subselected using grep.
}
\keyword{data}
