% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeNames.R
\docType{methods}
\name{changeNames}
\alias{changeNames}
\alias{changeNames,BioData-method}
\title{description of function changeNames}
\usage{
changeNames(x, what = "row", colname)
}
\arguments{
\item{x}{the BioData object}

\item{what}{which dimension names to change ('row' or 'col'; default='row')}

\item{colname}{a colname in the annotaion or samples table}
}
\description{
rename row- or colnames of the data tables
}
