% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUpGroup.R
\docType{methods}
\name{cleanUpGroup}
\alias{cleanUpGroup}
\alias{cleanUpGroup,BioData-method}
\title{description of function cleanUpGroup}
\usage{
cleanUpGroup(x, group, otherGroup, max_cells = 10, min_cells = 10, ...)
}
\arguments{
\item{x}{the BioData object}

\item{group}{the samples column to clean up}

\item{otherGroup}{the samples column to look for contaminations}

\item{max_cells}{how many cells to max use to create the predictive RF object (default = 10 [cells])}

\item{min_cells}{if two groups inside one cluster exceed min_cells one more cell group is created (defualt =1% of all cells)}

\item{...}{additional variables for the randomForest call}
}
\description{
Clean up intermixed groups
}
