% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_gene.R
\docType{methods}
\name{clusters_gene}
\alias{clusters_gene}
\alias{clusters_gene,BioData-method}
\title{description of function clusters_gene}
\arguments{
\item{dataObj}{the BioData object}

\item{clusterby}{cluster on raw data or MDS clustered data default="raw"}

\item{useGrouping}{do nothing and simply use tis grouping default=NULL}

\item{groups.n}{how many groups should be detected default= 3}

\item{ctype}{cluster type - either 'hierarchical clust' or 'kmeans' default = 'hierarchical clust'}

\item{onwhat}{this option has been kept for the Fluidigm data as there FACS data can also be used default = 'Expression'}

\item{cmethod}{the method to used with the hclust clustering (default = 'ward.D2')}

\item{name}{the name for the new grouping (default = 'auto_clusters_gene.1:n')}
}
\description{
Culters the data either based on the raw data values or any MDS data type and adds the grouping into the annotation table.
}

