% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonHeatmap.R
\docType{methods}
\name{CodonHeatmap}
\alias{CodonHeatmap}
\alias{CodonHeatmap,tRNAMINT-method}
\title{description of function heatmap}
\usage{
CodonHeatmap(x, colGroup, rowGroup = c("tRF.type.s."), norm.type = NULL,
  codon = NULL, fname = NULL, main = "Heatmap", collapse = NULL)
}
\arguments{
\item{x}{the tRNAMINT object}

\item{colGroup}{what do you want to have coloured at the top}

\item{rowGroup}{a list of annotaon table columns (default =  'tRF.type.s.')}

\item{norm.type}{any of 'all reads' 'all tRNA reads'  or 'Unnormalized'}

\item{codon}{which codon to focus on (a column in the annotation table)}

\item{fname}{the output filename}

\item{main}{the title string for the heatmap default="Heatmap"}

\item{collapse}{collapse the data before plotting; e.g. 
function(x) \{ collaps(x,what='row',fun = function(x) \{ sum( na.rm=TRUE) \} ) \}}
}
\description{
CodonHeatmap specific for tRNAMINT data. These heatmaps should be run on size ordered objects 
in order to visualize the read counts of the different tRNA fragments starting from the 5' over the center to the 3' and full length fragments.
}
