% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_collaps.R
\docType{methods}
\name{collaps}
\alias{collaps}
\alias{collaps,BioData-method}
\title{description of function collaps}
\arguments{
\item{dataObj}{the BioData}

\item{by}{collapsing method c('median','mean','sd','sum', or own function )}

\item{groupCol}{the sample names you want to group on}

\item{copy}{create a copy of the data (default=F)}
}
\description{
This function will collpase the data in the BioData to only contain one value
per sample group.
}
