% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse2codons.R
\docType{methods}
\name{collapse2codons}
\alias{collapse2codons}
\alias{collapse2codons,tRNAMINT-method}
\title{description of function collapse2codons}
\arguments{
\item{x}{the tRNAMINT object}

\item{sum.fun}{how to sum up the values for one sample default=function(x) { sum (x, na.rm=T)}}

\item{name}{the name of the collapsed object default="collapsed"}
}
\description{
This function collapses the data to the different codons and reports the sum.fun of all codons.
Single fragments can/will be counted multiple times.
}

