% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\docType{methods}
\name{colnames}
\alias{colnames}
\alias{colnames,BioData-method}
\title{warpper for colnames(x$dat)}
\arguments{
\item{x}{the BioData object}

\item{do.NULL}{see base::colnames (default= TRUE)}

\item{prefix}{see base::colnames  T default= "col"}
}
\description{
simple wrapper around colnames(x$dat)
}

