% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_4.R
\docType{methods}
\name{colors_4}
\alias{colors_4}
\alias{colors_4,BioData-method}
\title{description of function createRFgrouping_col}
\usage{
colors_4(x, name, colFunc = NULL, force = FALSE)
}
\arguments{
\item{x}{the BioData object}

\item{name}{the name of data column to colour}

\item{colFunc}{a colour function like default = function(x) {rainbow(x)}}

\item{force}{forcefully overwrite an existing color entry for this column (default FALSE)}
}
\description{
Create the colour information for a samples or annotation column
}

