% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_4.R
\name{colors_4}
\alias{colors_4}
\title{$name colors_4
$aliases colors_4,BioData-method
$rdname colors_4-methods
$docType methods
$description Create the colour information for a samples or annotation column
$param x the BioData object
$param name the name of data column to colour
$param colFunc a colour function like default = function(x) {rainbow(x)}
$title description of function createRFgrouping_col
$export}
\usage{
colors_4(x, name, colFunc = NULL)
}
\description{
$name colors_4
$aliases colors_4,BioData-method
$rdname colors_4-methods
$docType methods
$description Create the colour information for a samples or annotation column
$param x the BioData object
$param name the name of data column to colour
$param colFunc a colour function like default = function(x) {rainbow(x)}
$title description of function createRFgrouping_col
$export
}
