% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGroupings.R
\docType{methods}
\name{combineGroupings}
\alias{combineGroupings}
\alias{combineGroupings,BioData-method}
\title{description of function combineGroupings}
\usage{
combineGroupings(x, colGroups, new_name = "Merged Group 1",
  minCellOverlap = 5, minGroupOverlap = length(colGroups) - 2,
  minCellsInReturnGroup = NULL)
}
\arguments{
\item{x}{The BioData object}

\item{colGroups}{which groupings to combine (samples columns)}

\item{new_name}{the name of the new grouping (default="Merged Group 1" will replace existing information)}

\item{minCellOverlap}{while calculating the grouing overlap - how many cells need to overlap with each other default=5}

\item{minGroupOverlap}{while caulculating the overlap - how many groups should overlap in order to considder a cell overlapping default=length(colGroups)-2}

\item{minCellsInReturnGroup}{merge all groups with less than this number of cells into a ungrouped group default=1% of the bigges sname subgroup}
}
\description{
combine different grouping informations to create overlaps. 
Implemented to combine different subset RFclusterings into one including all features.
}
\details{
This function will calculate the overlap in group names for each cell to all other cells,
select the cells that match the requirement ( minCellOverlap and minGroupOverlap) and asigne the group id of the biggest group
to this cell.
The same algorythm will be applied for all cells that end up in small groups before the final ungrouped cells are assigned.
}
\examples{
combineGroupings(BioDataObj, colnames(BioDataObject$samples)[1:10], minCellOverlap = 5, minGroupOverlap = 8, minCellsInReturnGroup = 20)
}
