% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\docType{methods}
\name{convert_to}
\alias{convert_to}
\alias{convert_to,BioData-method}
\title{description of function convert_to}
\usage{
convert_to(x, type = c("MAST", "DESeq2", "scran"), species = NULL, ...)
}
\arguments{
\item{x}{the BioData object}

\item{type}{the type to convert to 'scran' should be the only working here....}

\item{species}{if converting to scran you need to tell me which species the data comes from in order to add the ensembol ids scan needs}

\item{...}{TEXT MISSING}
}
\description{
convert a biodata object into another object type
}
