% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_grouping.R
\docType{methods}
\name{copy_grouping}
\alias{copy_grouping}
\alias{copy_grouping,BioData-method}
\title{Copy a grouping (samples column) from one to another BioData object}
\usage{
copy_grouping(x, from, gname, newgname = NULL)
}
\arguments{
\item{x}{the traget object}

\item{from}{the source object}

\item{gname}{the (old) group name in the source object}

\item{newgname}{the (new) group name for the taget object default=NULL; use gname}
}
\description{
Copy a grouping from a (partial) BioData object
}
