% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_mds_objects.R
\docType{methods}
\name{copy_mds_objects}
\alias{copy_mds_objects}
\alias{copy_mds_objects,BioData-method}
\title{copy MDS structures and checking that the dimensions and order is OK.}
\usage{
copy_mds_objects(x, from, name = NULL, nameExt = NULL)
}
\arguments{
\item{x}{the BioData object the MDS tables should be copied to}

\item{from}{the BioData object that provides the MDS tables}

\item{name}{name the MDS you want to copy (default NULL copy all)}

\item{nameExt}{the attached name (default from$name)}
}
\description{
copy MDS struxtures from one BioData object to another.
By default the name of the otehr BioData object is combined to the MDS name
}
