% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_FindAllMarkers.R
\docType{methods}
\name{Cpp_FindAllMarkers}
\alias{Cpp_FindAllMarkers}
\alias{Cpp_FindAllMarkers,BioData-method}
\title{description of function Cpp_FindAllMarkers}
\usage{
Cpp_FindAllMarkers(x, condition, logfc.threshold = 1, minPct = 0.1,
  onlyPos = FALSE)
}
\arguments{
\item{x}{the BioData object}

\item{condition}{The grouping you wat to analyze}

\item{logfc.threshold}{seurat option used for speed up (default = 1)}

\item{minPct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function}

\item{onlyPos}{check only higher expression (default FALSE)}
}
\description{
Use Seurat's FindAllMarkers function
}
