% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBin.R
\docType{methods}
\name{CreateBin}
\alias{CreateBin}
\alias{CreateBin,BioData-method}
\title{Create a binned annotation column from numeric data}
\usage{
CreateBin(x, group = "nUMI", where = "sample", colFun = bluered)
}
\arguments{
\item{x}{the BioData object}

\item{group}{the group (defaule = 'nUMI'}

\item{where}{in the samples (sample) or annotation (gene) data frame (sample)}

\item{colFun}{colory function default=  bluered}
}
\description{
Bin the UMI data into 13 bins for plotting and define a blue <- red color gradient
}
