% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDuplicates.R
\docType{methods}
\name{createDuplicates}
\alias{createDuplicates}
\alias{createDuplicates,BioData-method}
\title{description of function createDuplicates}
\arguments{
\item{x}{a SingleCells object}

\item{group}{which group should be the basis to create duplicates (only inter group duplicates will be created)}

\item{randomCellsFrac}{how many random cell should be created default=.1}
}
\description{
Create artificial duplicates using the raw data stored in the object.
}
