% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfCluster_row.R
\docType{methods}
\name{createRFgrouping_row}
\alias{createRFgrouping_row}
\alias{createRFgrouping_row,BioData-method}
\title{description of function createRFgrouping_row}
\usage{
createRFgrouping_row(x, RFname = "notExisting", k = 10,
  single_res_row = paste("RFgrouping", RFname), colFunc = NULL)
}
\arguments{
\item{x}{the BioData object}

\item{RFname}{the name of the RFclust.SGE object in the BioData object. This object has to be populized with data!}

\item{k}{the number of wanted groups ( default = 10)}

\item{single_res_row}{the new column in the samples table default= paste('RFgrouping', RFname)}

\item{colFunc}{a function giving the colours back for the grouping (gets the amount of groups) default = function(x){rainbow(x)}}
}
\description{
Create a sample grouping data from one RFclust.SGE object
}
