% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStats.R
\docType{methods}
\name{createStats}
\alias{createStats}
\alias{createStats,NGSexpressionSet-method}
\title{description of function createStats}
\arguments{
\item{x}{the NGSexpressionSet}

\item{condition}{the grouping column in the samples data.frame}

\item{files}{write the statistics tables (FALSE)}

\item{A}{the first component to analyze (Group A)}

\item{B}{the second component to analyze (Group B)}
}
\value{
the NGSexpressionSet with a set of ststs tables
}
\description{
calculate staistics on all possible groupings using the DEseq nbinomTest test Both

together create the group 'A vs. B'
}
