% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2biodata.R
\docType{methods}
\name{dataframe2biodata}
\alias{dataframe2biodata}
\alias{dataframe2biodata,BioData-method}
\title{description of function dataframe2biodata}
\arguments{
\item{x}{the data table}

\item{name}{the name of the new BioData object}
}
\description{
This function creates a BioData object from a data frame only assuming, that the first column of the data frame
contains the row names. The sample table will be constructed from the colnames(x)[-1] vector.
}

