% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_grouping.R
\docType{methods}
\name{define_grouping}
\alias{define_grouping}
\alias{define_grouping,BioData-method}
\title{description of function define_grouping}
\usage{
define_grouping(x, basedOn, colname, newNames, what = "col")
}
\arguments{
\item{x}{The BioData object}

\item{basedOn}{the old grouping name}

\item{colname}{the new grouping name}

\item{newNames}{the list "with new_group_name" = c( "old Group Names" ) structure}

\item{what}{create a sample ('col') or annotation ('row') grouping}
}
\description{
This will create a new sample or annotation column based on an old grouping and a list.
}
\examples{
# data is a BioData object with a sample group 'oldG' and the levels c( 1, 2, 3, 4)
define_grouping( data, 'oldG', 'newGrouping', list( 'A' = c( 1,4), B= c(2,3)), 'col')
# will create a 'newGrouping' column in the $samples table with the entries A and B
}
