% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineGOIs.R
\docType{methods}
\name{defineGOIs}
\alias{defineGOIs}
\alias{defineGOIs,cellexalvr-method}
\title{description of function defineGOIs}
\usage{
defineGOIs(x, name, genes, lables = NULL, gene_col = NULL)
}
\arguments{
\item{x}{A BioData object}

\item{name}{the name of the GIO list (eg TFs or epigenetic)}

\item{genes}{a list of gene symbols that match to the $dat rownames}

\item{lables}{a list of lables for the GIO column (default NULL)}

\item{gene_col}{which column in the annotation data contains the IDs you give me
default == rownames(x$dat)}
}
\description{
Allows the user to define (G)enes (O)f (I)nterest lists in the object
}
