% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineHeatmapColors.R
\docType{methods}
\name{defineHeatmapColors}
\alias{defineHeatmapColors}
\alias{defineHeatmapColors,BioData-method}
\title{description of function defineHeatmapColors}
\arguments{
\item{x}{the BioData object}

\item{melted}{the merged data object with the Expression column that should be colored}

\item{colrs}{and optional colors vector( gray + bluered for data and rainbow for samples)}

\item{lowest}{the lowest color in the heatmap (gray by default)}
}
\value{
a list with the modified merged table and the colors vector
}
\description{
uses ggplot2 to plot heatmaps
}

