% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineHeatmapColors.R
\name{defineHeatmapColors}
\alias{defineHeatmapColors}
\title{$name defineHeatmapColors
$aliases defineHeatmapColors,BioData-method
$rdname defineHeatmapColors-methods
$docType methods
$description  uses ggplot2 to plot heatmaps
$param merged the merged data object with the Expression column that should be colored
$param colrs and optional colors vector( gray + bluered for data and rainbow for samples)
$param lowest the lowest color in teh heatmap (gray by default)
$title description of function gg.heatmap.list
$return a list with the modified merged table and the colors vector
$export}
\usage{
defineHeatmapColors(x, melted, ..., colrs = NULL, lowest = "gray")
}
\description{
$name defineHeatmapColors
$aliases defineHeatmapColors,BioData-method
$rdname defineHeatmapColors-methods
$docType methods
$description  uses ggplot2 to plot heatmaps
$param merged the merged data object with the Expression column that should be colored
$param colrs and optional colors vector( gray + bluered for data and rainbow for samples)
$param lowest the lowest color in teh heatmap (gray by default)
$title description of function gg.heatmap.list
$return a list with the modified merged table and the colors vector
$export
}
