% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\docType{methods}
\name{DimReduction}
\alias{DimReduction}
\alias{DimReduction,BioData-method}
\title{initial dimensional reduction step based on PCA}
\arguments{
\item{x}{the BioData object}

\item{genes}{create the dim red for the genes dimension (rows) default=FALSE}

\item{n}{how many eigenvectors to collapse to default=100}

\item{method}{which method ('auto', 'irlba', 'bpca')}

\item{force}{re-produced the dataset even if it already exists default=FALSE}
}
\value{
the name of the result object in the usedObj list.
}
\description{
Create the initial dimension reduction dataset (PCA n =100)
}
