% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.data.R
\docType{methods}
\name{export.data}
\alias{export.data}
\alias{export.data,BioData-method}
\title{description of function write.data}
\usage{
export.data(x)
}
\arguments{
\item{x}{the BioData object}
}
\description{
write the BioData data, annotation and samples tables to disk
}
