% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCodonInformation.R
\docType{methods}
\name{extractCodonInformation}
\alias{extractCodonInformation}
\alias{extractCodonInformation,tRNAMINT-method}
\title{description of function extractCodonInformation}
\usage{
extractCodonInformation(x,
  col = "Sequence.locations.in.tRNA.space..comma.deliminated.")
}
\arguments{
\item{x}{the tRNAMINT object}

\item{col}{the column conatining the codon information default= "Sequence.locations.in.tRNA.space..comma.deliminated."}
}
\description{
Converts the MINT tRNA annoation strings into codon information stored in the annotation.
}

