% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGenes.R
\docType{methods}
\name{filterGenes}
\alias{filterGenes}
\alias{filterGenes,BioData-method}
\title{description of function filterGenes}
\arguments{
\item{x}{the BioData object}

\item{genes}{a list of pattern matches to the egnes you want to remove}
}
\description{
Filter out genes following a different expression profile
}
\details{
Mitocondrial genes are expressed not from the diploid genome, 
but from multiple copies of the mitochondrial genome.
A similar problem arises with the ribosomal genes which are 
expressed from multiple locis in the diploig genome and therfore also spike at much higher 
levels than the diploid genes.

I am not 100% sure about the ribosomal genes, but the mitochondial ones should definitely be filtered out!
}
\examples{
\dontrun{
## merged is a BioData object with mouse data
filterGenes( merged, c('^mt-', 'Rp[ls]'))
## merged is a BioData object with human data
filterGenes( merged, c('^MT', 'RP[LS]'))
}
}
