% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDuplicates.R
\docType{methods}
\name{findDuplicates}
\alias{findDuplicates}
\alias{findDuplicates,BioData-method}
\title{description of function findDuplicates}
\arguments{
\item{x}{a SingleCells object}

\item{group}{a grouing focusing on the differences between groups. Too many groups would be bad here!}

\item{randomCellsFrac}{how many cells should be created as random duplicates default=.01}

\item{iterations}{each iteration takes significant amount of time default= 10}
}
\description{
This function is a re-implementation of the DuplicateFinder logics
}
\details{
In contrast to the publiushed function this will use the re-sampling based normalization the BioData obejct provides.
In addition this function uses the BioData internal VarGenes function.

The results are totally untested and a lot of additional work is likely necessary.
}
