% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlapWithCells.R
\docType{methods}
\name{findOverlapWithCells}
\alias{findOverlapWithCells}
\alias{findOverlapWithCells,BioData-method}
\title{description of function findOverlapWithCells}
\arguments{
\item{x}{the BioData obejct}

\item{cells}{the cells to find similarity for (do not take too view)}

\item{cname}{the column name to store the similarity values in default="SimilarityToTargetCells"}

\item{iter}{the amount of kmean runs for this analysis default=100}

\item{group.n}{the amount of kmeans groups to be created}
}
\description{
PCA and kmeans based similarity finder
}
\details{
Based on a PCA representation of the BioData object this function runs 'iter'
kmeans clusterings and collects the mean fraction of 'cells' overlapping with each other cell.

The function is returung this fraction of overlap for each cells and the higher the values the better the overlap.
}
