% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_4_rf.R
\docType{methods}
\name{fit_4_rf}
\alias{fit_4_rf}
\alias{fit_4_rf,BioData-method}
\title{Reduce memory footprint and -1 values from a BioData object}
\usage{
fit_4_rf(x, copy = TRUE)
}
\arguments{
\item{x}{The BioData object}

\item{copy}{Return a copy of the object default= TRUE}
}
\description{
fit_4_rf removes raw and zscored data and all -1 values from the dat object
}
