% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.numeric.R
\docType{methods}
\name{force.numeric}
\alias{force.numeric}
\alias{force.numeric,BioData-method}
\title{description of function force.numeric}
\arguments{
\item{x}{the BioData object}
}
\description{
make sure, that all variables in the data table are numeric. Add NAs if necessary.
}
