% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{fun_df2list}
\alias{fun_df2list}
\title{change the data.frame into list}
\usage{
fun_df2list(
  input_df,
  input_name_regex = NULL,
  name_col = "term",
  value_col = "gene"
)
}
\arguments{
\item{input_df}{data.frame}

\item{input_name_regex}{regex pattern to search in data.frame column data}

\item{name_col}{which column to search(list names in results)}

\item{value_col}{the values(list values in results)}
}
\value{
list
}
\description{
change the data.frame into list
}
\examples{
 tmp = fun_df2list("NF",db_hallmark)
}
