% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{fun_gene2symbol}
\alias{fun_gene2symbol}
\title{convert gene data into symbol}
\usage{
fun_gene2symbol(
  input_genes,
  from = "alias",
  gene_db_df = biodata::gene_alias2symbol,
  alias_not_found_keep_or_drop = c("drop", "keep")
)
}
\arguments{
\item{input_genes}{input_genes}

\item{from}{from type}

\item{gene_db_df}{the db used for gene mapping}

\item{alias_not_found_keep_or_drop}{whether keep notfound data}
}
\value{
data.frame
}
\description{
convert gene data into symbol
}
\examples{
fun_gene2symbol(c("ACTB","CT",'afdfs',"MT1"))
}
