% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.genes.cor.to.R
\docType{methods}
\name{get.genes.cor.to}
\alias{get.genes.cor.to}
\alias{get.genes.cor.to,BioData-method}
\title{description of function get.genes.cor.to}
\arguments{
\item{x}{the BioData object}

\item{gname}{the gene name of interst}

\item{output}{the VR outpath}
}
\description{
VR function that exports all genes correlating to the input gene name
}
