% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsembl.R
\docType{methods}
\name{getEnsembl}
\alias{getEnsembl}
\alias{getEnsembl,BioData-method}
\title{description of function getEnsembl}
\usage{
getEnsembl(term, species = "mMus")
}
\arguments{
\item{term}{the GO accession number to use e.g. 'GO:0007049' CellCycle}

\item{species}{the organism to use either default= 'mMus' or 'Hs'}
}
\value{
A vector of ENSEMBL gene ids
}
\description{
Function copied from https://github.com/PMBio/scLVM/blob/master/R/scLVM/R/util.R
is a simple helper to get GO annotations from either org.Mm.egGO2EG or org.Hs.egGO2EG, both BioConductor packages.
}
