% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\docType{methods}
\name{getGeneInfo}
\alias{getGeneInfo}
\alias{getGeneInfo,BioData-method}
\title{description of function getEnsembl}
\usage{
libraray(BioData)
CellCycle_gene_symbols <- getGeneInfo(
			"GO:0007049", species = 'mMus', from= 'go_id', from_tab='go' 
)
CellCycle_ensembl_ids <- getGeneInfo(
			"GO:0007049", species = "Hs", from = "go_id", from_tab = "go", 
			what='ensembl_id', what_tab="ensembl")
Complement_and_coagulation_cascades_gene_symbols <- getGeneInfo(
			'04610', species = 'Hs', from = 'path_id', what= 'symbol' )
}
\arguments{
\item{x}{the query string or a vector of query strings}

\item{species}{either 'mMus', 'Hs' or a AnnotationDbi object}

\item{from}{the column that corresponds to the x values}

\item{from_tab}{if the column is defined in more thanone table the specififc table name}

\item{what}{which column do you want to get (only one)}

\item{what_tab}{analogue to the from_tab the table name containing the colmn if more than one table contains the info}
}
\value{
A vector of ids
}
\description{
getGeneInfo is a poverful function, that uses the annotation packages RSQLite tables to select information.
}
