% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesExpressedHigherThanExpected.R
\docType{methods}
\name{getGenesExpressedHigherThanExpected}
\alias{getGenesExpressedHigherThanExpected}
\alias{getGenesExpressedHigherThanExpected,BioData-method}
\title{description of function getGenesExpressedHigherThanExpected}
\arguments{
\item{x}{a SingleCells object}

\item{group}{a samples grouping used to subset the original object}

\item{ref}{the correlating samples group (default nUMI)}

\item{n}{get the top interesting genes default= 300 (in each group!)}
}
\description{
description of function getGenesExpressedHigherThanExpected
}
\details{
After excessive clustering and really good grouping of the cells this function can be apllied to single cells dataset
to identify genes that are expressed to a higher level than what the correlation to the UMI count can explain.

This strategy is hily questionable, but results in quite interesting gene lists. No normalization is required for this function to work.
It is always applied to the not normalized data.
}
