% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNgenesFromMultiGroupStats.R
\docType{methods}
\name{getNgenesFromMultiGroupStats}
\alias{getNgenesFromMultiGroupStats}
\alias{getNgenesFromMultiGroupStats,BioData-method}
\title{description of function getNgenesFromMultiGroupStats}
\arguments{
\item{all_markers}{the data frame containg all results (columns 'gene', 'cluster' are important)}

\item{num.sig}{how many genes to return (default=1000)}
}
\description{
return the top significant genes in an even spacing for all analyzed goups.
}
