% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.R
\docType{methods}
\name{getSymbols}
\alias{getSymbols}
\alias{getSymbols,BioData-method}
\title{description of function getSymbols}
\usage{
getSymbols(ensIds, species = "mMus")
}
\arguments{
\item{ensIds}{the ids to check}

\item{species}{the organism to use either default= 'mMus' or 'Hs'}
}
\description{
Function copied from https://github.com/PMBio/scLVM/blob/master/R/scLVM/R/util.R
is a simple helper to get Gene Symbols from ENSEMBL ids using either org.Mm.egENSEMBL2EG or org.Hs.egENSEMBL2EG, both BioConductor packages.
}
