% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_intersect_order.R
\docType{methods}
\name{group_intersect_order}
\alias{group_intersect_order}
\alias{group_intersect_order,BioData-method}
\title{description of function group_intersect_order}
\usage{
group_intersect_order(x, group, otherGroup)
}
\arguments{
\item{x}{the BioData object}

\item{group}{the group to be reordered}

\item{otherGroup}{the leading grouping}
}
\description{
use a second ordering to reorder a given sample grouping
}
