% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.boxplot.R
\docType{methods}
\name{groups.boxplot}
\alias{groups.boxplot}
\alias{groups.boxplot,BioData-method}
\title{description of function groups.boxplot}
\arguments{
\item{x}{the ExpresionSet object}

\item{groupCol}{the column in the samples table that contains the sample grouping information}

\item{groupRow}{the column in the annotation table hat contains the gene grouping information}

\item{svg}{create the single boxplots as svg or (default png) files}

\item{fname}{the filename extension files are created by paste(x$outpath,fname,<GroupID>,"_boxplot")}

\item{Collapse}{if set will lead to a collapse of the sample groups into one value per gene. Supports all \code{\link{collaps}} by options.}
}
\description{
This function can be used to get an overview of the different gene level groups in

an BioData It uses the linux montage command to merge all different boxplots

into one figure
}
