% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_loom2BioData.R, R/readLoomFile.R
\docType{methods}
\name{H5Anno2df}
\alias{H5Anno2df}
\alias{H5Anno2df,BioData-method}
\alias{H5Anno2df,cellexalvrR-method}
\title{description of function H5Anno2df}
\usage{
H5Anno2df(x, slotName, namecol = NULL, onlyStrings = FALSE)

H5Anno2df(x, slotName, namecol = NULL, onlyStrings = FALSE)
}
\arguments{
\item{x}{the H5 object}

\item{slotName}{the H5 entity tro convert to a data.frame}

\item{namecol}{the (optional) rownames column for the data}

\item{onlyStrings}{return only columns that not only contain numbers (default FALSE)}
}
\description{
convert a H5 annotation (any name) table to a data table

convert a H5 annotation (any name) table to a data table
}
