% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\docType{methods}
\name{heatmap}
\alias{heatmap}
\alias{heatmap,tRNAMINT-method}
\title{description of function heatmap}
\usage{
heatmap(x, colGroup, norm.type = NULL, codon = NULL, fname = NULL,
  main = "Heatmap")
}
\arguments{
\item{x}{the tRNAMINT object}

\item{colGroup}{what do you want to have coloured at the top}

\item{norm.type}{any of 'all reads' 'all tRNA reads'  or 'Unnormalized'}

\item{codon}{which codon to focus on (a column in the annotation table)}

\item{fname}{the output filename}

\item{main}{the title string for the heatmap default="Heatmap"}
}
\description{
heatmap specific for tRNAMINT data. These heatmaps should be run on size ordered objects 
in order to visualize the read counts of the different tRNA fragments starting from the 5' over the center to the 3' and full length fragments.
}
