% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatPlotTimelines2D.R
\docType{methods}
\name{heatPlotTimelines2D}
\alias{heatPlotTimelines2D}
\alias{heatPlotTimelines2D,BioData-method}
\title{description of function heatPlotTimelines2D}
\arguments{
\item{obj}{TEXT MISSING}

\item{times}{TEXT MISSING}

\item{mds}{TEXT MISSING}

\item{grouping}{TEXT MISSING}

\item{genes}{TEXT MISSING}

\item{outpath}{TEXT MISSING default=NULL}

\item{sumN}{TEXT MISSING default=10}

\item{x}{TEXT MISSING default=1}

\item{y}{TEXT MISSING default=2}
}
\description{
Plot a summed up gene expression on top of a 2D dimension reduction and
}
\details{
Plot summed up gene expression on top of a 2D dimension reduction a 
sample grouping and a timeline.
}
