% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_files.R
\docType{data}
\name{human.CellSurface}
\alias{human.CellSurface}
\title{A simple list containing the human ENTREZ.gene.symbols from 'A Mass Spectrometric-Derived Cell Surface Protein Atlas' plos 2015}
\format{
vector
}
\usage{
human.CellSurface
}
\description{
This list is used to crete a CellSurface object
}
\keyword{data}
