% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_data_files.R
\docType{data}
\name{human.tfs}
\alias{human.tfs}
\title{A list of human transcription factors.}
\format{A vector 1468 in length:almdiR_Aurroa

\describe{
  \item{tf}{transcription factor}
  ...
}}
\source{
\url{http://bioinfo.life.hust.edu.cn/AnimalTFDB/}
}
\usage{
human.tfs
}
\description{
A list of human transcription factors. Used when generating TF networks
}
\keyword{datasets}
