% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyMarkerGenes.R
\docType{methods}
\name{IdentifyMarkerGenes}
\alias{IdentifyMarkerGenes}
\alias{IdentifyMarkerGenes,BioData-method}
\title{description of function IdentifyMarkerGenes}
\usage{
IdentifyMarkerGenes(x, gname, settings = list(), names = NULL)
}
\arguments{
\item{x}{the BioData object}

\item{gname}{the samples column name to group on.}

\item{settings}{a list of slurm parameters to use to run a script (optional)}

\item{names}{a vector of group name values to compare against all others (default NULL == use all)}
}
\description{
This method uses whichever stats function was selected for this class using the createStats function.
But it compares one group versus all other groups to find marker genes for this group only.
This function will add length(group) new stats tables.
}
