% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_CellexalVR_selection.R
\docType{methods}
\name{load_CellexalVR_selection}
\alias{load_CellexalVR_selection}
\alias{load_CellexalVR_selection,BioData-method}
\title{description of function load_CellexalVR_selection}
\arguments{
\item{x}{the BioData object}

\item{file}{the cellexal selection<id>.txt file}

\item{gname}{the sample group name this selection should get in the BioData object}
}
\description{
load a selection.txt file created during a CellexalVR run
}
