% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021_loadObj.R
\docType{methods}
\name{loadObj}
\alias{loadObj}
\alias{loadObj,BioData-method}
\title{description of function loadObj}
\usage{
loadObj(file = NULL)
}
\arguments{
\item{file}{the infile}
}
\description{
This function loads the a BioData file and returns the data R object int them.
 The function is mainly to keep the scripts simpler. There is no magic involved.
}

