% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\docType{methods}
\name{logThis}
\alias{logThis}
\alias{logThis,BioData-method}
\title{description of function log}
\usage{
logThis(x)
}
\arguments{
\item{x}{the biodata object}
}
\description{
calculate the log for all data +1
}
