% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make3D4Obj.R
\docType{methods}
\name{Make3D4obj}
\alias{Make3D4obj}
\alias{Make3D4obj,BioData-method}
\title{description of function Make3D4obj}
\arguments{
\item{x}{the BioData object}

\item{group}{the grouping value (either colname from the samples table or a rowname from the data table}

\item{mds.type}{the mds type to use (names from names(x$usedObj$MDS)) default = PCA}

\item{cex}{define the size of the strings default=0.5}

\item{colFunc}{the color function if the color is not already defined using colors_4()  default= function(x){rainbow(x)}}

\item{cut}{this has to be true for genes, as the samples are then binned into 10 expression groups each containing the same number of samples default=F}

\item{names}{if true not dots, but the name in the samples table is plotted in 3D default=F}

\item{opath}{create a webgl representation of the plot in a html page in this path (default = NULL)}

\item{main}{the title of the plot (default ='')}

\item{genes}{use gene level MDS data not sample level (default = FALSE)}

\item{plotType}{choose one [1,2] and check whether you like it ;-) default=1}

\item{size}{the size of the 3D points default = 3.0}

\item{green}{single cell normalization looses gene expression values. Display the cells with lost expression in green or black default=FALSE}
}
\description{
This function  created a MDS 3D plot with an inbuilt background describing the data colors
}
