% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,BioData-method}
\title{description of function merge}
\usage{
merge(x, objects = list(), anno_merge, anno_uniq)
}
\arguments{
\item{x}{TEXT MISSING}

\item{objects}{TEXT MISSING default=list()}

\item{anno_merge}{TEXT MISSING}

\item{anno_uniq}{TEXT MISSING}
}
\description{
merges two BioData objects
}
