% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cells.R
\docType{methods}
\name{merge_cells}
\alias{merge_cells}
\alias{merge_cells,BioData-method}
\title{description of function merge_cells}
\usage{
merge_cells(dataObj, group, mergeNcells = 10, gID = NULL, by = c("median",
  "mean", "sd", "sum", "var"), copy = TRUE)
}
\arguments{
\item{dataObj}{the BioData object}

\item{group}{the original grouping you want to merge on}

\item{mergeNcells}{merge mergeNcells into one summary cell (default =10)}

\item{gID}{merge only cells in a spicific ID list of this group; Say the group as entries 'A', 'B', and 'C', but you only want to merge cells on groupID 'A' and 'B' you need to specifiy ('A', 'B') here (default =NULL)}

\item{by}{by collapsing method c('median','mean','sd','sum', or own function )}

\item{copy}{create a copy of the R6 object instead of changing the real thing (default=TRUE)}
}
\description{
merge cells to get summary values of median values or anything you want
}
