% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineHeatmapColors.R
\docType{methods}
\name{minValueExpr}
\alias{minValueExpr}
\alias{minValueExpr,BioData-method}
\title{description of function gg.heatmap.list}
\usage{
minValueExpr(x)
}
\arguments{
\item{x}{data object}
}
\value{
a list with the modified merged table and the colors vector
}
\description{
just get me the min value for the object
}
