% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineHeatmapColors.R
\name{minValueExpr}
\alias{minValueExpr}
\title{$name minValueExpr
$aliases minValueExpr,BioData-method
$rdname minValueExpr-methods
$docType methods
$description just get me the min value for the object
$param x data object
$title description of function gg.heatmap.list
$return a list with the modified merged table and the colors vector
$export}
\usage{
minValueExpr(x)
}
\description{
$name minValueExpr
$aliases minValueExpr,BioData-method
$rdname minValueExpr-methods
$docType methods
$description just get me the min value for the object
$param x data object
$title description of function gg.heatmap.list
$return a list with the modified merged table and the colors vector
$export
}
