% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_files.R
\docType{data}
\name{mouse.tfs}
\alias{mouse.tfs}
\title{A list of mouse transcription factors.}
\format{A vector 1357 in length:
\describe{
  \item{tf}{transcription factor}
  ...
}}
\source{
\url{http://bioinfo.life.hust.edu.cn/AnimalTFDB/}
}
\usage{
mouse.tfs
}
\description{
A list of mouse transcription factors. Used when generating TF networks
}
\keyword{datasets}
