% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,BioData-method}
\alias{normalize,SingleCells-method}
\alias{normalize,MicroArray-method}
\title{description of function normalize}
\usage{
\S4method{normalize}{BioData}(object, ..., readCounts = NULL,
  to_gene_length = FALSE, geneLengthCol = "transcriptLength",
  force = FALSE, name = NULL)

\S4method{normalize}{SingleCells}(object, ..., reads = 600,
  force = FALSE, name = NULL)

\S4method{normalize}{MicroArray}(object, to = NULL, ..., name = NULL)
}
\arguments{
\item{readCounts}{The number of reads from each bam file or another value you want to normalize the data to}

\item{to_gene_length}{FALSE whether or not to normalize the data to gene length}

\item{geneLengthCol}{the column in the annotation data.frame to (in addition) normalize the genes to (e.g. trancript length)}

\item{force}{re-normalize this object (default FALSE)}

\item{name}{the new name of the object (deafule old name + normalized)}

\item{reads}{the required read depth}

\item{to}{a numeric vector to normalize the samples to. Has to have the same length as are columns in the data table}

\item{x}{The BioData object (NGS expression data, not single cells)}

\item{forec}{replace old norm data (FALSE)}

\item{x}{The SingleCellsNGS object}

\item{name}{the name of the new object}

\item{x}{the BioData object}
}
\value{
the normalized data set (original data stored in NGS$raw

the normalized data set (original data stored in slot 'raw'
}
\description{
normalize the expression data (sample wise)

normalize the expression data by subsampling as described in PMID 24531970

constructor that has to be implemented for a generic BioData
This generic version was meant for array data and I have not had the need nor time to implement this part.
}
