% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\docType{methods}
\name{normalize_R_buggy}
\alias{normalize_R_buggy}
\alias{normalize_R_buggy,BioData-method}
\title{normalize a SingleCells::BioData::R6 object}
\arguments{
\item{x}{The SingleCells::BioData::R6 object}

\item{reads}{the required read depth}

\item{name}{the name of the new object}

\item{force}{re-normalize this object (default FALSE)}
}
\value{
the normalized data set (original data stored in slot 'raw'
}
\description{
normalize the expression data by subsampling as described in PMID 24531970
}
