% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyLogPage.R
\docType{methods}
\name{ontologyLogPage}
\alias{ontologyLogPage}
\alias{ontologyLogPage,BioData-method}
\title{description of function ontologyLogPage}
\usage{
ontologyLogPage(x, genes, ontology = "BP", topNodes = 10,
  GOfname = "GOgenes.csv", ...)
}
\arguments{
\item{x}{the BioData object}

\item{genes}{a list of gene symbols (IMPORTANT)}

\item{ontology}{which GO ontology to choose from (default = "BP")}

\item{topNodes}{how many pathways to return (default 10)}

\item{GOfname}{file to export the results to (default GOgenes.csv)}

\item{...}{unused}
}
\description{
creates the GO analysis for a gene list and puts it into the report.
}
