% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialTests.R
\docType{methods}
\name{PartialTests}
\alias{PartialTests}
\alias{PartialTests,BioData-method}
\title{description of function PartialTests}
\usage{
PartialTests(
  obj,
  groupA,
  groupB,
  pcut = 1e-05,
  logfc.threshold = 0.1,
  minPct = 0.1
)
}
\arguments{
\item{obj}{The BioData object}

\item{groupA}{the outer grouping column name}

\item{groupB}{the inner grouping column name}

\item{pcut}{p cutoff value for gene selection default=1e-5}

\item{logfc.threshold}{Cpp test option default= .1}

\item{minPct}{Cpp test option default= .1}
}
\description{
Run tests firs subselecting the BioData object on the outer grouping and calculating tests on the inner grouping.
}
\details{
PartialTests should take two column names and apply stat tests on sub-samples using group1 
and test for differentials in group2.
The stat results are saved inside the original R6 object
All differential genes are returned as list.
}
