% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2col_barplot.R
\docType{methods}
\name{plot2col_barplot}
\alias{plot2col_barplot}
\alias{plot2col_barplot,BioData-method}
\title{description of function plot2col_barplot}
\arguments{
\item{x}{the BioData object}

\item{n1}{the samples column for the x axis}

\item{n2}{the samples column for the y axis and the coloring of the bars#' @name plot2col_barplot}

\item{file}{the outfile for this plot default=NULL}

\item{svg}{create a svg file default=F (png)}

\item{pdf}{create a pdf default=F (png)}

\item{xScaleFactor}{rescale width for very long factor names take higher value ( default=10 )}

\item{X11type}{a type for the plot devices default='cairo'}
}
\description{
Plot n1 vs n2 columns as colored barplot to show the distribution of n2 inside of n1
}
