% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_velocyto.R
\docType{methods}
\name{plot_velocyto}
\alias{plot_velocyto}
\alias{plot_velocyto,BioData-method}
\title{description of function plot_velocyto}
\usage{
plot_velocyto(x, velo, translation, group, mds_name = "Expression PCA",
  ret = list(), ofile = NULL, ...)
}
\arguments{
\item{x}{The BioData object}

\item{velo}{the velocyto read reads}

\item{translation}{a two columns table with 'velo' and 'BioData' sample names (translation table)}

\item{group}{the group to be plotted}

\item{mds_name}{the mds to plot on (default "Expression PCA")}

\item{ret}{an optional list gotten from the last plot_velocyto run (default=list())}

\item{ofile}{an optional outfile for the figure (pdf) default= NULL}

\item{...}{unused at the moment}
}
\value{
a list of velocyto.R results (reusable)
}
\description{
use a velocyto input and plot the velocities on one of the MDS projections
}
