% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.codonUsage.R
\docType{methods}
\name{plotCodonUsage}
\alias{plotCodonUsage}
\alias{plotCodonUsage,BioData-method}
\title{description of function plotCodonUsage}
\arguments{
\item{x}{the tRNAMINT object}

\item{sname}{the sample you want to depict}

\item{codons}{the codon list to use default=all}

\item{min_reads}{use all tRNA fragments with at least that many reads for the sample default=1}

\item{tRF.reliability}{MINT reports an exclusive and ambiguous set of tRNA fragments state one of them or use both default=NULL == both}

\item{tRF.type}{MINT reports different types of tRNA fragments (3'-half, 5'-half, ...) default=NULL == use all}

\item{main}{the figure title default="Pie Chart"}

\item{color}{a vector of color information to use for the codons default= rainbow}

\item{fname}{the filename to plot to as png (default use x11 as plotting device)}
}
\description{
plots the codon usage for one sample
The function uses sampleCodonUsage() to extract this information from for the sample
}

