% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.legend.R
\docType{methods}
\name{plotLegend}
\alias{plotLegend}
\alias{plotLegend,BioData-method}
\title{description of function plot.legend}
\usage{
plotLegend(
  x,
  colname,
  file = NULL,
  svg = F,
  pdf = F,
  col = NULL,
  X11type = "cairo",
  family = "Helvetica"
)
}
\arguments{
\item{x}{the StefansExpressionSet object}

\item{colname}{the name of a column in the annotation or samples tables that should be described in the legend}

\item{file}{the optional outfile}

\item{svg}{create a svg file default=F}

\item{pdf}{create a pdf file default=F}

\item{col}{a vector of color names default=NULL}

\item{family}{the grDevices font family (default 'Helvetica')}
}
\description{
Creates the legend for one sample or annotation variable using either the inbuilt 
color or an external one in the col option. If a file is given, the data will be plotted into a file. 
If neither pdf of svg is true the data will be plotted to a png file. File endings will always be added - so do not do that yourself.
}
