% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\docType{methods}
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,BioData-method}
\title{description of function plotMDS}
\arguments{
\item{obj}{The BioData object}

\item{mds}{The name of the MDS to plot}

\item{group}{Which sample grouping(s) to plot onm the MDS}

\item{genes}{Which genes to color the MDS with}

\item{x}{MDS dimension on the x axis default=1}

\item{y}{MDS dimension on the y axis default=2}

\item{green}{show the gene expression lost during normalization in green default=F}

\item{pdf}{use pdf instead of png output (default TRUE)}

\item{width}{width of the plots (default pdf 10)}

\item{height}{height of the plots (default pdf 10)}

\item{...}{additional plot arguments for the 2D dr plot}
}
\description{
Create 2D plts of any MDS and a list of feature plots
}
