% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivesOnly.R
\docType{methods}
\name{positivesOnly}
\alias{positivesOnly}
\alias{positivesOnly,BioData-method}
\title{description of function positivesOnly}
\arguments{
\item{x}{the SingleCells object}

\item{what}{which slot to get default='dat'}
}
\description{
get a stored Matrix data set without the normalization drop out information
}
