% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\docType{methods}
\name{preprocess}
\alias{preprocess}
\alias{preprocess,BioData-method}
\title{description of function preprocess}
\arguments{
\item{x}{the BioData}

\item{condition}{the column in the samples table to use as grouing variable
   The condition will be used as formular paste('~', condition) 
and can therfore contain more complex function like 'batch + condition'}
}
\description{
create the count dataset and normalize it following DEseq standard

function used by DEseq Do not use
}
