% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeTest3D.R
\docType{methods}
\name{pseudotimeTest3D}
\alias{pseudotimeTest3D}
\alias{pseudotimeTest3D,BioData-method}
\title{description of function pseudotimeTest3D}
\arguments{
\item{x}{a BioData object (preferably a SingleCells object)}

\item{a}{the first time defining vector (e.g. a MDS dimension)}

\item{b}{the second time defining vector (e.g. a MDS dimension)}

\item{grouping}{a samples grouping to color the Pseudotime plot.}

\item{outpath}{a outpath for the files  default=x$outpath}

\item{n}{how many genes to plot on both sides default=100}

\item{plotGenes}{additional default genes to plot default=NULL}

\item{smooth}{roling smoothing window size (default 100)}

\item{invert}{invert the time before usage ( default=FALSE)}

\item{cleanFolder}{remove all files from the outpout folder before creating new ones (default FALSE)}

\item{plotType}{( 'png', 'png_high_res', 'pdf' )}

\item{summaryPlot}{the name of the summary plot file ( default NULL no summary plot)}
}
\description{
Calculate and plot pseudotime relevant information.
}
\details{
A simple pseudotime analysis based on a linear model, that can create the pseudotime from the two varibales 'a' and 'b'

fit = loess( b~a )

The fitted values will be used as pseudotime and the pseudotiume is plotted 
versus the two defininge variables to the file outpath/Pseudotime.png using the grouping color.
Pseudotime is depicted as bluered line indicating the orientation blue -> red.

The pseudotime will be added into the samples table and the correlation to 
the pseudotime will be added to the annotation table (both replacing old data!).

The function returns the top/bottom (n) correlating genes.

The rolling mean values that are summing up only expression values from one group 
are colored in the group colour all others are black.
}
