% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwd.R
\docType{methods}
\name{pwd}
\alias{pwd}
\title{description of function pwd}
\usage{
pwd(a)
}
\value{
A string containing the working directory
}
\description{
uses the linux pwd command to determin the working directory
}
