% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_of_fit.R
\docType{methods}
\name{quality_of_fit}
\alias{quality_of_fit}
\alias{quality_of_fit,BioData-method}
\title{description of function quality_of_fit}
\arguments{
\item{obj}{The BioData object}

\item{what}{cells or genes - which grouping do you want to check?}

\item{col}{group name as one column in the samples table}
}
\description{
Calculates a quality of fit
}
