% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLoomFile.R
\docType{methods}
\name{readLoomFile}
\alias{readLoomFile}
\alias{readLoomFile,BioData-method}
\title{description of function readLoomFile}
\arguments{
\item{x}{the loom file}
}
\description{
read a loom file
}
