% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTo.R
\docType{methods}
\name{reduceTo}
\alias{reduceTo}
\alias{reduceTo,BioData-method}
\title{description of function reduceTo}
\arguments{
\item{x}{the NGScollation object}

\item{what}{reduce to samples or row ids default='row'}

\item{to}{select these names default=NULL}
}
\description{
The main reduction function can drop both samples and genes using the colnames / rownames of the data tables
}
