% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNeg.R
\docType{methods}
\name{removeNeg}
\alias{removeNeg}
\alias{removeNeg,BioData-method}
\title{description of function removeNeg}
\arguments{
\item{x}{teh SingleCells object}

\item{from}{which slot to remove the data from (dat or zscored) default='dat'}
}
\description{
removes -1 values from a SingleCells object
}
