% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew.R
\docType{methods}
\name{renew}
\alias{renew}
\alias{renew,BioData-method}
\title{description of function renew}
\arguments{
\item{x}{the object you want to update}
}
\description{
update the class definition by re-creating the instance
}

