% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_reorder.obj.R
\name{reorder.genes}
\alias{reorder.genes}
\title{$name reorder.genes
$aliases reorder.genes,BioData-method
$rdname reorder.genes-methods
$docType methods
$description this function reorderes the BioData object based on a column in the samples table (e.g. for plotting)
$param x the BioData object
$param column the samples column to reorder on
$title description of function remove.genes
$export}
\usage{
reorder.genes(x, column)
}
\description{
$name reorder.genes
$aliases reorder.genes,BioData-method
$rdname reorder.genes-methods
$docType methods
$description this function reorderes the BioData object based on a column in the samples table (e.g. for plotting)
$param x the BioData object
$param column the samples column to reorder on
$title description of function remove.genes
$export
}
