% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_reorder.obj.R
\name{reorder.samples}
\alias{reorder.samples}
\title{$name reorder.samples
$aliases reorder.samples,BioData-method
$rdname reorder.samples-methods
$docType methods
$description this function reorderes the BioData object based on a column in the annotation table (e.g. for plotting)
$param x the BioData object
$param column the annotation column to reorder on
$title description of function remove.genes
$export}
\usage{
reorder.samples(x, column)
}
\description{
$name reorder.samples
$aliases reorder.samples,BioData-method
$rdname reorder.samples-methods
$docType methods
$description this function reorderes the BioData object based on a column in the annotation table (e.g. for plotting)
$param x the BioData object
$param column the annotation column to reorder on
$title description of function remove.genes
$export
}
