% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_grouping.R
\docType{methods}
\name{reorder_grouping}
\alias{reorder_grouping}
\alias{reorder_grouping,BioData-method}
\title{description of function reorder_grouping}
\usage{
reorder_grouping(obj, group, new_order, what = "col")
}
\arguments{
\item{obj}{the BioDat object}

\item{group}{the sample group to re-order}

\item{new_order}{the new order (ordered factor levels)}

\item{what}{samples or annotation column to re-order? (default 'col')}
}
\description{
Reorder a grouping after some checks that the id string is suitable to do this
}
