% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_on_wilcox_results.R
\docType{methods}
\name{reorder_on_wilcox_results}
\alias{reorder_on_wilcox_results}
\alias{reorder_on_wilcox_results,BioData-method}
\title{description of function reorder_on_wilcox_results}
\arguments{
\item{x}{the BioData object}

\item{statsN}{stats table to use}

\item{column}{samples column to change}

\item{cutoff}{p value cut of for the wilcox tables default=1e-8}
}
\description{
uses a wilcox result table gene overlap between the groups to identify close groups.
}
