% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfCluster_row.R
\docType{methods}
\name{rfCluster_row}
\alias{rfCluster_row}
\alias{'rfCluster_row,BioData-method}
\title{rfCluster_row}
\usage{
rfCluster_row(x, rep = 5, SGE = F, email = "none", k = 16, slice = 30,
  subset = 200, nforest = 500, ntree = 500, name = "RFclust")
}
\arguments{
\item{x}{the single cells ngs object}

\item{rep}{how many repetitions for the random forest grouping should be run (default = 5)}

\item{SGE}{whether to use the sun grid engine to calculate the rf grouping}

\item{email}{your email to use together with the SGE option}

\item{k}{the numer of expected clusters (metter more than to view)}

\item{slice}{how many processes should be started for each random forest clustering (default = 30)}

\item{subset}{how many cells should be randomly selected for the unsupervised clustering (default = 200)}

\item{nforest}{the numer of forests to grow for each rep (defualt = 500)}

\item{ntree}{the numer of trees per forest (default = 500)}

\item{name}{if you want to run multiple RFclusterings on e.g. using different input genes you need to specify a name (default ='RFclust')}

\item{bestColname}{the column name to store the results in}
}
\value{
a SingleCellsNGS object including the results and storing the RF object in the usedObj list (bestColname)
}
\description{
This fucntion uses the RFclust.SGE to create fandomForest based unsupervised clusters on a subset of the data.

Default is on 200 cells using all (provided) genes with 500 forests and 500 trees per forest for 5 repetitions.

You are asked to give a k numer of expected clusters (better too many than too little), classifies the total

data using the 5 different unsupervised runs and all cluster ids from these runs are merged into the final cluster id.

This <summaryCol> will be part of the return objects samples table, together with a <usefulCol> where

all clusters with less than 10 cells have been merged into the 'gr. 0'.

The final results will be reported as new columns in the samples table containing the 'name'
}
