% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyLogPage.R
\docType{methods}
\name{rmdLink}
\alias{rmdLink}
\alias{rmdLink,BioData-method}
\title{easily create an Rmd link that opens in a new window.}
\arguments{
\item{name}{the displayed text}

\item{link}{the link address}

\item{lineEnd}{add a line end at the end of every entry (default =T)}
}
\description{
creates a linke in the structure [<name>](<link><name>){target='blank'}\n
}
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyLogPage.R
\docType{methods}
\name{rmdLink}
\alias{rmdLink}
\alias{rmdLink,BioData-method}
\title{easily create an Rmd link that opens in a new window.}
\arguments{
\item{name}{the displayed text}

\item{link}{the link address}

\item{lineEnd}{add a line end at the end of every entry (default =T)}
}
\description{
creates a linke in the structure [<name>](<link><name>){target='blank'}
}
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyLogPage.R
\docType{methods}
\name{rmdLink}
\alias{rmdLink}
\alias{rmdLink,BioData-method}
\title{easily create an Rmd link that opens in a new window.}
\arguments{
\item{name}{the displayed text}

\item{link}{the link address}

\item{lineEnd}{add a line end at the end of every entry (default =T)}
}
\description{
creates a linke in the structure [<name>](<link><name>){target='blank'}
}
