% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\docType{methods}
\name{rownames}
\alias{rownames}
\alias{rownames,BioData-method}
\title{wrapper for rownames(x$dat)}
\arguments{
\item{x}{TEXT MISSING}

\item{do.NULL}{see base::colnames (default= TRUE)}

\item{prefix}{see base::colnames  T default= "row"}
}
\description{
wrapper for rownames(x$dat)
}

