% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStats_inThread.R
\docType{methods}
\name{runStats_inThread}
\alias{runStats_inThread}
\alias{runStats_inThread,BioData-method}
\title{description of function runStats_inThread}
\usage{
runStats_inThread(x, condition, files = F, A = NULL, B = NULL,
  covariates = NULL, form = NULL, settings = NULL)
}
\arguments{
\item{x}{the BioData object}

\item{condition}{the samples column to run stats for}

\item{files}{whether to print the stats output to file ( default =F - depricated)}

\item{A}{if not all conditions should be used condition A}

\item{B}{if not all conditions should be used condition B}

\item{covariates}{should covariates be used - name them here}

\item{form}{a specific formualr to use? State it here}
}
\description{
create a script that calculates the stats in a new thread; run the script and collect the data after the script has finished
}

