% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCodonUsage.R
\docType{methods}
\name{sampleCodonUsage}
\alias{sampleCodonUsage}
\alias{sampleCodonUsage,tRNAMINT-method}
\title{description of function sampleCodonUsage}
\arguments{
\item{x}{the tRNAMINT object}

\item{sname}{the sample you want to depict}

\item{codons}{the codon list to use default=all}

\item{min_reads}{use all tRNA fragments with at least that many reads for the sample default=1}

\item{tRF.reliability}{MINT reports an exclusive and ambiguous set of tRNA fragments state one of them or use both default=NULL == both}

\item{tRF.type}{MINT reports different types of tRNA fragments (3'-half, 5'-half, ...) default=NULL == use all}
}
\description{
Calculates the sum of tRNA fragments for one sample that come from a specific tRNA for one codon.
}
