% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratVarGenes.R
\docType{methods}
\name{SeuratVarGenes}
\alias{SeuratVarGenes}
\alias{SeuratVarGenes,BioData-method}
\title{description of function SeuratVarGenes}
\usage{
SeuratVarGenes(x, minVar = 1)
}
\arguments{
\item{x}{The BioData object}

\item{minVar}{the min scaled varianze (default=1)}
}
\description{
Pathetic function to use the Seurat FindVariableFeatures function
}
