% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class.BioData.R
\docType{methods}
\name{show}
\alias{show}
\title{the BioData show function}
\usage{
\S4method{show}{BioData}(object)
}
\arguments{
\item{object}{the BioData object}
}
\description{
the BioData show function
}
