% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class.BioData.R
\docType{class}
\name{SingleCells}
\alias{SingleCells}
\title{Class interface for single cell data}
\format{\code{\link{R6Class}} object.}
\usage{
SingleCells
}
\value{
Object of \code{\link{R6Class}} to store single cell data.
}
\description{
Class interface for single cell data
}
\section{Fields}{

\describe{
\item{\code{data}}{the numerical data as data.frame}

\item{\code{samples}}{the sample annotation as data.frame}

\item{\code{annotation}}{the row annotation as data.frame}

\item{\code{usedObj}}{a multi purpose list to store whichever ananlyis results do not fit in the stats list}

\item{\code{stats}}{all stats with one result for each data row}
}}

\examples{
set.seed(1)
dat = data.frame( matrix(rnorm(1000),ncol=10) ) 
colnames(dat) <- paste('Sample', 1:10)
rownames(dat) <- paste( 'gene', 1:100)
samples <- data.frame(SampleID = 1:10, sname = colnames(dat) )
annotation <- data.frame( GeneID = paste( 'gene', 1:100), Start= 101:200 )
x <- SingleCells$new( cbind(annotation,dat), Samples=samples, name="testObject",namecol='sname', outpath = "" )
}
\keyword{NGS}
\keyword{cell,}
\keyword{single}
