% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split4Geo.R
\docType{methods}
\name{Split4Geo}
\alias{Split4Geo}
\alias{Split4Geo,BioData-method}
\title{description of function Split4Geo}
\usage{
Split4Geo(x, opath = x$outpath)
}
\arguments{
\item{x}{the BioData object}

\item{opath}{the outpath if different from x$outpath}
}
\description{
This function will store each column in the data slot table as separate file rbound to the rownames of the data object.
The name of the outfile will be the column name + .xls and it will be a tab separated table.
An additional md5sums.txt file will be created containing all md5sums for the exported files.
}
