% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLite_ExpressionSummary.R
\docType{methods}
\name{SQLite_ExpressionSummary}
\alias{SQLite_ExpressionSummary}
\alias{SQLite_ExpressionSummary,BioData-method}
\title{description of function SQLite_ExpressionSummary}
\usage{
SQLite_ExpressionSummary(fname)
}
\arguments{
\item{fname}{the file to the sqlite db}
}
\description{
Connect to a sqlite database and retrieve a gene read count summary for all the stored genes.
}
