% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLite_SampleSummary.R
\docType{methods}
\name{SQLite_SampleSummary}
\alias{SQLite_SampleSummary}
\alias{SQLite_SampleSummary,BioData-method}
\title{description of function SQLite_SampleSummary}
\usage{
SQLite_SampleSummary(
  fname,
  cells = list(table = "samples", rev = "sample_id", name = "sname")
)
}
\arguments{
\item{fname}{the file to the sqlite db}

\item{cells}{details about the cell storage table default list( 'table' = 'samples', 'rev' = 'sample_id', 'name' = 'sname')}
}
\description{
Connect to a sqlite database and retrieve a sample read count summary for all the stored genes.
}
