% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_table.R
\docType{methods}
\name{stat_table}
\alias{stat_table}
\alias{stat_table,tRNAMINT-method}
\title{description of function stat_table}
\usage{
stat_table(l, padjMethod = "BH")
}
\arguments{
\item{l}{the list to convert}

\item{padjMethod}{the method name for the p.adjust call (default = 'BH')}
}
\description{
converts a list of tables one table adding the list name as first column
}
