% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_data_files.R
\docType{data}
\name{TestData}
\alias{TestData}
\title{merged the two smallest Mouse Tabular muris datasets facs_Aorta_TabulaMuris and droplet_Heart_TabulaMuris}
\format{BioData}
\usage{
TestData
}
\description{
two TabulaMuris mouse datasets merged
}
\keyword{datasets}
