% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\docType{methods}
\name{transpose}
\alias{transpose}
\alias{transpose,02_transpose-method}
\title{description of function t}
\arguments{
\item{x}{the BioData object}
}
\description{
transpose the object DELETING all stats and usedObj!
}
