% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRF.typeHeatmap.R
\docType{methods}
\name{tRF.typeHeatmap}
\alias{tRF.typeHeatmap}
\alias{tRF.typeHeatmap,tRNAMINT-method}
\title{description of function heatmap}
\arguments{
\item{x}{the tRNAMINT object}

\item{colGroup}{what do you want to have coloured at the top}

\item{norm.type}{any of 'all reads' 'all tRNA reads'  or 'Unnormalized'}

\item{tRF.type}{which tRF.type to focus on (a column in the annotation table)}

\item{fname}{the output filename}

\item{main}{the title string for the heatmap default="Heatmap"}

\item{fun}{collapse the data before plotting; e.g.}

\item{z.score}{z.score the data after the summary (default = F)}

\item{brks}{how many shades of heatmap colors ( default 10 )
function(x) \{ collaps(x,what='row',group='frag.type.and.length', fun = function(x) \{ sum( x, na.rm=TRUE) \} ) \}}
}
\description{
tRF.typeHeatmap specific for tRNAMINT data. These heatmaps should be run on size ordered objects 
in order to visualize the read counts of the different tRNA fragments starting from the 5' over the center to the 3' and full length fragments.
}
