% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_stats.R
\docType{methods}
\name{tRNA_stats}
\alias{tRNA_stats}
\alias{tRNA_stats,tRNAMINT-method}
\title{description of function tRNA_stats}
\usage{
tRNA_stats(x, acol, scol, norm.type = NULL, codon = NULL,
  fun = function(x) {     sum(x, na.rm = TRUE) })
}
\arguments{
\item{x}{the tRNAMINT object}

\item{acol}{the statistics will be calculated for every entry in this column}

\item{scol}{column containing the sample grouping (two groups are required)}

\item{norm.type}{a tRNAMINT object normally contains 3 or more different norm datasets per sample.
This value is used as a patter to select one normalization mode.}

\item{codon}{the tRNAMINT object can extract codon information from the fragments using the extractCodonInformation() function.
If you give one codon column name here (x$usedObj$Codons) the stats will be calculated for this codon only}

\item{fun}{multiple fragments are summed up for this analysis; this is the summing function: default=function(x) { sum(x, na.rm=TRUE) }}
}
\description{
This function calculates the tRNA statistics as described in XYZ
}

