% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useInbuiltGOIlists.R
\docType{methods}
\name{useInbuiltGOIlists}
\alias{useInbuiltGOIlists}
\alias{useInbuiltGOIlists,BioData-method}
\title{description of function useInbuiltGOIlists}
\usage{
useInbuiltGOIlists(x, name, ...)
}
\arguments{
\item{x}{A BioData object}

\item{name}{the name of the inbuilt list to use ( either 'TFs' or 'epigenetic' for now)}

\item{gene_col}{which column in the annotation data contains the Gene.Symbols? 
The inbuilt groups are Gene.Symbol based.}
}
\description{
An easy function to register the inbuilt (G)enes (O)f (I)nterest lists 'TFs' and

'epigenetic' are supported at the moment
}
