% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_vioplot.R
\docType{methods}
\name{vioplot}
\alias{vioplot}
\alias{vioplot,Rscexv-method}
\title{description of function vioplot}
\usage{
vioplot(x, ..., range = 1.5, h = NULL, ylim = NULL, names = NULL,
  horizontal = FALSE, col = "magenta", border = "black", lty = 1,
  lwd = 1, rectCol = "black", colMed = "white", pchMed = 19, at,
  add = FALSE, wex = 1, drawRect = TRUE, main = NULL,
  cex.axis = 1, neg = NULL)
}
\arguments{
\item{x}{a numerix vector}

\item{...}{additional numerix vectors to plot}

\item{range}{TEXT MISSING default= 1.5}

\item{h}{TEXT MISSING default= NULL}

\item{ylim}{TEXT MISSING default= NULL}

\item{names}{TEXT MISSING default= NULL}

\item{vioplot}{TEXT MISSING}
}
\description{
the vioplot function patched to allow different colors for the different violines.
}
\details{
This function is a hack of the official vioplot function to allow multiple colors of the violines.
}
