% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.score.R
\docType{methods}
\name{z.score}
\alias{z.score}
\alias{z.score,BioData-method}
\title{description of function z.score}
\usage{
z.score(m)
}
\arguments{
\item{m}{the matrix of column = samples and rows = genes or a BioData object}
}
\value{
the z scored matrix / BioData object
}
\description{
z score the matrix
}

