% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_stat.R
\docType{methods}
\name{add_to_stat}
\alias{add_to_stat}
\alias{add_to_stat,BioData-method}
\title{Add a stsitics table to the $stats slot}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{The BioDat object}

\item{stat}{the stats table}

\item{name}{the name for this stats table}
}
\description{
Adds a stst table to the object
}
\keyword{datasets}
