% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BioData.R
\docType{methods}
\name{as_BioData}
\alias{as_BioData}
\title{description of function as_BioData}
\arguments{
\item{dat}{the counts list you get from Rsubread::featureCounts()}
}
\description{
create a NGSexpressionSet from a counts list object
}

