% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BioData.R
\docType{methods}
\name{as_BioData}
\alias{as_BioData}
\alias{as_BioData,list-method}
\alias{as_BioData,matrix-method}
\alias{as_BioData,data.frame-method}
\alias{as_BioData,cellexalvrR-method}
\alias{as_BioData,seurat-method}
\alias{as_BioData,character-method}
\title{create a BioData from an other object}
\usage{
\S4method{as_BioData}{list}(dat)

\S4method{as_BioData}{matrix}(dat)

\S4method{as_BioData}{data.frame}(dat)

\S4method{as_BioData}{cellexalvrR}(dat)

\S4method{as_BioData}{seurat}(dat)

\S4method{as_BioData}{character}(dat, minUMI = 100, minGexpr = NULL)
}
\arguments{
\item{dat}{a list coming for Rsubread scanning of NGS data}

\item{minUMI}{the minimum UMI count in order to read a cell (default=100)}

\item{minGexpr}{the minimum UMI count for a gene to be included (default NULL - translates into 1)}

\item{...}{source object specififc options}

\item{dat}{a matrix object with cells in columns and genes in rows}

\item{dat}{a cellexlvrR object}

\item{dat}{a cellexlvrR object}

\item{dat}{a seurat object}

\item{dat}{a sqlite database as created by the cellexalvrR::export2cellexalvr() function}
}
\description{
create a BioData from an other object
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: create a BioData::R6 object from a Rsubread result list

\item \code{matrix}: convert a matrix to a BioData::R6 object

\item \code{data.frame}: convert a data.frame to a BioData::R6 object

\item \code{cellexalvrR}: convert a cellexalvrR object to a SingleCells::BioData::R6 object

\item \code{seurat}: convert a seurat object into a SingleCells::BioData::R6 object

\item \code{character}: character path or sqlite3 database name
}}

