% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_CellexalvrR.R
\docType{methods}
\name{as_cellexalvr}
\alias{as_cellexalvr}
\alias{as_cellexalvr,BioData-method}
\title{description of function as_cellexalvr}
\usage{
as_cellexalvr(x, outpath, specie, meta.cell.groups,
  meta.genes.groups = NULL, userGroups = NULL, minGene = 1)
}
\arguments{
\item{x}{The BioData object}

\item{outpath}{the outpath for the VR data}

\item{specie}{which specie the data comes from (mouse or human)}

\item{meta.cell.groups}{which BioData samples groups should become available in the Attributes coloring mode in VR}

\item{meta.genes.groups}{at the moment saver to leave NULL}

\item{userGroups}{which sample columns should become accessible as userGroupings in VR to re-color the graphs}

\item{minGene}{VR has a problem with 0 level genes, hence these need to be removes, but you can also set a higher cut off (default=1)}
}
\description{
export a BioData as cellexalVR folder that can be loaded into the CellexalVR VR platform
}
