% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestGrouping.R
\docType{methods}
\name{bestGrouping}
\alias{bestGrouping}
\alias{'bestGrouping,BioData-method}
\title{description of function randomForest}
\usage{
bestGrouping(x, group, bestColname = "QualifiedGrouping", cutoff = 0.5)
}
\arguments{
\item{x}{the single cells ngs object}

\item{group}{a vector of sample columns that should be checked (the most complex is used only)}

\item{bestColname}{the column name to store the best grouping in}

\item{cutoff}{the cutoff percentage where all groups showing less than this percentacge of remapped samples are dropped}
}
\value{
a distRF object to be analyzed by pamNew
}
\description{
The function is using a randomForest classifier with 2000 trees to classify the given data using the given grooping

All groups that fail to be prediceted using the random forest are deemed ungrouped.

All groups where less than 50 percent of the total samples geting classified as being from that group fail.
}
