% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata-package.R
\docType{package}
\name{biodata-package}
\alias{biodata}
\alias{biodata-package}
\title{biodata: Datasets for the life sciences}
\description{
Datasets and example analyses for the life sciences..
}
\author{
\strong{Maintainer}: Nathan Brouwer \email{brouwernt@gmail.com}

}
\keyword{internal}
