% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_grouping.R
\docType{methods}
\name{define_grouping}
\alias{define_grouping}
\alias{define_grouping,BioData-method}
\title{description of function define_grouping}
\usage{
define_grouping(x, basedOn, colname, newNames, what = "col")
}
\arguments{
\item{x}{The BioData object}

\item{basedOn}{the old grouping name}

\item{colname}{the new grouping name}

\item{newNames}{the list "with new_group_name" = c( "old Group Names" ) structure}

\item{what}{create a sample ('col') or annotation ('row') grouping}
}
\description{
This will create a new sample or annotation column based on an old grouping and a list.
}
