% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_PartialTest_Obj.R
\docType{methods}
\name{Get_PartialTest_Obj}
\alias{Get_PartialTest_Obj}
\alias{Get_PartialTest_Obj,BioData-method}
\title{description of function Get_PartialTest_Obj}
\usage{
Get_PartialTest_Obj(obj, groupA, groupB, pcut = 1e-05,
  logfc.threshold = 0.1, minPct = 0.1)
}
\arguments{
\item{obj}{The BioData object}

\item{groupA}{the outer grouping column name}

\item{groupB}{the inner grouping column name}

\item{pcut}{the cutoff p value default=1e-5}

\item{logfc.threshold}{Cpp test option default= .1}

\item{minPct}{Cpp test option default= .1}
}
\description{
Process the 2 group comparison data and return a summary object
}
