% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesExpressedHigherThanExpected.R
\docType{methods}
\name{getGenesExpressedHigherThanExpected}
\alias{getGenesExpressedHigherThanExpected}
\alias{getGenesExpressedHigherThanExpected,BioData-method}
\title{description of function getGenesExpressedHigherThanExpected}
\arguments{
\item{x}{a SingleCells object}

\item{group}{a samples grouping used to subset the original object}

\item{ref}{the correlating samples group (default nUMI)}

\item{n}{get the top interesting genes default= 300 (in each group!)}
}
\description{
My personal VarGenes function.
}
\details{
To identify variable genes is almost the first thing that every single cell analysis strategy does.
The existing VarGene function identify many highly expressed genes like 
ribosomal genes or mitochondrial genes if they have not been removed before.

This normally produces problems in the downstream analysis as they normally correlate with the cell cycle which
most of the times is not the main interesting cellular process.

Here we try to identify genes that are correlated less to the UMI count than would be expected.
The function estimates the normal correlation values for each gene count and selects for genes that
are correlated to the UMI count less than other genes with the same fraction of cells expressing them.

This has proven to especially pick genes at the lower expression range like 
chromatin remodeling proteins or transcription factors.

No normalization is required for this function to work.
It is always applied to the not normalized data.
}
