% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,BioData-method}
\title{merge 2 or more BioData objects}
\usage{
merge(x, objects = list())
}
\arguments{
\item{x}{the first BioData object}

\item{objects}{a list of all others to merge to}
}
\description{
merges two or more BioData objects
}
\examples{
objects # is a list of BioDat objects having the same gene annotation type (e.g. ENSEMBL IDs!)
data <- objects[[1]]
objects[[1]] = NULL
merged <- merge ( data, objects)
}

