% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_reorder.obj.R
\docType{methods}
\name{reorder.genes}
\alias{reorder.genes}
\alias{reorder.genes,BioData-method}
\title{description of function remove.genes}
\arguments{
\item{x}{the BioData object}

\item{column}{the samples column to reorder on}
}
\description{
this function reorderes the BioData object based on a column in the samples table (e.g. for plotting)
}
