% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022_saveObj.R
\docType{methods}
\name{saveObj}
\alias{saveObj}
\alias{saveObj,BioData-method}
\title{description of function saveObj}
\usage{
saveObj(data, file = NULL)
}
\arguments{
\item{data}{the BioData object}

\item{file}{the outfile default=paste(str_replace_all(data$name, "\\s+", "_" ), '.RData',sep='')}
}
\description{
This function saves the object either as analysis.RData or norm_data.RData if the analysi.RData has not been produced before
}
