% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratVarGenes.R
\docType{methods}
\name{SeuratVarGenes}
\alias{SeuratVarGenes}
\alias{SeuratVarGenes,BioData-method}
\title{description of function SeuratVarGenes}
\usage{
SeuratVarGenes(x, n = 300)
}
\arguments{
\item{x}{The BioData object}

\item{n}{the maount of genes you want (default 300)}
}
\description{
Pathetic function to use the Seurat FindVariableFeatures function
}
