% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,BioData-method}
\title{description of function show}
\usage{
\S4method{show}{BioData}(object)
}
\arguments{
\item{x}{the BioData object}
}
\value{
nothing
}
\description{
show the BioData
}

