% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLite_SampleSummary.R
\docType{methods}
\name{SQLite_SampleSummary}
\alias{SQLite_SampleSummary}
\alias{SQLite_SampleSummary,BioData-method}
\title{description of function SQLite_SampleSummary}
\usage{
SQLite_SampleSummary(fname)
}
\arguments{
\item{fname}{the file to the sqlite db}
}
\description{
Connect to a sqlite database and retrieve a sample read count summary for all the stored genes.
}
