% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class.BioData.R
\docType{data}
\name{tRNAMINT}
\alias{tRNAMINT}
\title{tRNAMINT}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tRNAMINT
}
\description{
An R6 class to visualize Expression data.
}
\section{Slots}{

\describe{
\item{\code{objects}}{a data.frame containing the expression values for each gene x sample (gene = row)}

\item{\code{samples}}{a data.frame describing the columnanmes in the data column}

\item{\code{annotation}}{a data.frame describing the rownames in the data rows}

\item{\code{outpath}}{the default outpath for the plots and tables from this package}

\item{\code{name}}{the name for this package (all filesnames contain that)}

\item{\code{zscored}}{genes are normalized?}

\item{\code{snorm}}{samples normalized?}

\item{\code{usedObj}}{here a set of used and probably lateron important objects can be saved. Be very carful using any of them!}
}}

\keyword{datasets}
